Shadow Cat VMF & Guru Seat Condition (#132)
* Edited shadow cat vmf condition

Player now need to go to the Vending Machine Factory and having seen at least once the cat. We can't unfortunately make a better trigger linked to variables or Switches since none of them are used for checking that.

* Attempt to fix Guru Seat by changing trigger

Changed trigger to be like the ones of the angel nurses of Deep Dreams since they use similar conditions. If this still don't work, I guess we could plan to just ask a coordinate trigger.

* Update cube_guru_seat.json