Added the translation of some references
The The Red Card badge's description is a line from the video game A Hat in Time, and was taken from each of the game's official localisation.
The lines from Someday were taken from the Spanish and Russian translations, and the Dragon Ball reference from the net.

Remaining: French localisation, removal of Ena's badges.