Batch 122 - Badge Shuffle (#804)
* Update some badges

- Remove the Four Seasons badge by rollaby on request
- Replace rollaby's Spa Day badge by an art made by Sekiranun on request
- Standardise JL's name to Jellyfish Legerdemain and WikiOdd to wiki0dd
- Made the Pinwheel Countryside badge be able to be triggered in its second map (thank you Nulsdodage for the report)

* Remove Ena's badges on request

Remove 21 badges and alter the credits of 5 badges. (lang edits for those and the Four Seasons badge to come)
Collaboration are still kept, just with her uncredited.

* Add 41+2 badges for batch 122

32 Yume 2kki (2 of them being replacements of removed badges), 5 Collective Unconscious, 3 .flow, 2 Uneven Dream, and 1 Dream Genie badge.

Yet to do: translation of lines in French, translations of Snatcher's line and reimplemented conditions, removal of lang lines of requested badges.

* Fix the pt lang file and add the translation of the condition lines

* Added the translation of some references

The The Red Card badge's description is a line from the video game A Hat in Time, and was taken from each of the game's official localisation.
The lines from Someday were taken from the Spanish and Russian translations, and the Dragon Ball reference from the net.

Remaining: French localisation, removal of Ena's badges.

* Remove the requested badges from the lang files

* Translate some lines + minor adjustements

* Update fr.json

* Add checkboxes for the menu types badge