Add 12 badges for batch 143 (#920)
* Add 12 badges for batch 143

* Fix the Spirit Capital badges

The map ID has been changed in the recent 710 update, which was not reflected here

* Fix the B-Sides badge

The map of Downfall Garden B was updated, meaning that its trigger also needed to be updated

* Update the condition the Monochromatic Void badge

See ynolocations#167

* Update the location and condition of the Gothic Beauty badge

The effect got moved from White Desert to Body Part Exhibition, so both needed to be updated to make sense

* Change the condition of the Expert of Disguise badge

Can now be unlocked by completing the two first pages of the mask menu.
Removed the triggers tied to the chaser masks due to no longer being available.

* Fix the Angler's Depths trigger for the Triangulated Skeleton badge

The X coordinates were outdated so it wouldn't be triggered as intended when playing

* Update the condition text of the Scarlet Extinction badge