Update radon_noble_gas_district.json (#985)
* Update radon_noble_gas_district.json

* Update Ticks' crediting

* Attempt to fix the NASU badges

During the Game Over, the switch for the cheat code gets reset, but the variable to track the score only gets reset when starting a new game, which seems to lead to the possibility of getting the NASU badges with a lucky timing while using the cheat code. Since it's not possible to check switches to determine if the player is playing NASU (reset before the variable change) and that some variables are reused before, I thought of making the trigger be the display of the red background picture before the Game Over (only occurs while the cheat code switch is properly set, so no way of a thing breaking on that). It will likely not work due to being nonsense and I'll have to revert this commit but yay 👍

* Update nasu_ruby.json