Temporarily remove the Cosmic Horrors badge
Said area was made inaccessible by mistake in 0.3.2; it was fixed in 0.3.3, but the fix was unitentionally reverted in 0.3.4. Should be fixed in 0.4.0, but for now since the badge can't be obtained it's best to remove it.

Conditions and lang entries kept, as they should not change and shouldn't matter