Add 16 badges for batch 182 (#1150)
* Add 16 badges for batch 182

* Temporarily remove the Cosmic Horrors badge

Said area was made inaccessible by mistake in 0.3.2; it was fixed in 0.3.3, but the fix was unitentionally reverted in 0.3.4. Should be fixed in 0.4.0, but for now since the badge can't be obtained it's best to remove it.

Conditions and lang entries kept, as they should not change and shouldn't matter

* Simplify the Sunken Girl badge

Remove the no effect trigger (too specific while also being for other effects normally), Chainsaw trigger (already part of the blood trigger), Railroad Crossing trigger (confusing since it doesn't change her reaction), Fairy/Trombone trigger (requires hitting reconnect)

* Undummy the Amber Fields and Crimson Dusk badge

* Fix the Retaliation badge's description

* Update ja.json

* Update cleaner_ending.json

* Add data for 1 more batch 183 badge

* Add data for 2 more batch 183 badges

* Update night_visit_deterioration.json

* Update moonlit_pharmacy_society.json

* Add data for 4 more batch 183 badges